ciao-controller: move common mac address generation func to utils
Since mac address generation is needed for both the datastore and main
package, move these funcs into a new "utils" package, which is included
by both. Add some unit testing for the new utils package.
